lab-3-iot-hub/functions/requirements.txt
azure-functions
pymysql